Separate needed functions into utils/*.mjs to make them reusable
